blb-141 slight refactoring before dasher implementation
