+++ Assets/Scripts/Entities/Enemies.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 92ec93004b1c3044cb935d18ca763098
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
