+++ Assets/Scripts/Entities/Enemies/Dasher.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a4af50f36cf2a694991d9ecd6288232e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
