+++ Assets/Scripts/FiniteStateMachine/States/EnemyStates/DasherStates/DasherAttackState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 117bf6be0e60d03408fd62820cc77d19
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
