+++ Assets/Scripts/FiniteStateMachine/States/EnemyStates/DasherStates/DasherDeathState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1bae9f25a1f46b04f8c3261dff137ae3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
