+++ Assets/Scripts/FiniteStateMachine/States/EnemyStates/DasherStates/DasherIdleState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4879c1138f5e6d7498d96d80a7ecb138
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
