+++ Assets/Scripts/FiniteStateMachine/States/EnemyStates/DasherStates/DasherPatrolState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 92b2ad0ba1c63c4478f3bc1035e13c28
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
