+++ Assets/Scripts/FiniteStateMachine/States/EnemyStates/EnemyAlertedState.cs
@@ -29,7 +29,7 @@ public class EnemyAlertedState : EnemyState
 
         if (Time.time >= startTime + alertTime)
         {
-            stateMachine.ChangeState(enemy.ChaseTargetState);
+            stateMachine.ChangeState(enemy.AgroState);
             return;
         }
