+++ Assets/Scripts/FiniteStateMachine/States/EnemyStates/EnemyCCState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e2977bbe426def3439844ba7d961c270
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
