+++ Assets/Scripts/FiniteStateMachine/States/EnemyStates/EnemyDeathState.cs
@@ -1,0 +1,34 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+// not implemented yet
+public class EnemyDeathState : EnemyState
+{
+    public EnemyDeathState(Enemy enemy, FiniteStateMachine stateMachine) : base(enemy, stateMachine) { }
+
+    public override void Enter()
+    {
+        base.Enter();
+    }
+
+    public override void Exit()
+    {
+        base.Exit();
+    }
+
+    public override void LogicUpdate()
+    {
+        base.LogicUpdate();
+    }
+
+    public override void DoChecks()
+    {
+        base.DoChecks();
+    }
+
+    public override void PhysicsUpdate()
+    {
+        base.PhysicsUpdate();
+    }
+}
