+++ Assets/Scripts/FiniteStateMachine/States/EnemyStates/EnemyDeathState.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 446b616c4f130b34fbb11d220ffc502c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
